projects/rag-qa-system/requirements.txt
# RAG Q&A System - Requirements
# Created: 2025-11-23 (Day 27)
# Install with: pip install -r requirements.txt

# Core dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
openai==2.8.1
python-dotenv==1.2.1

# PDF parsing
pymupdf4llm==0.2.3

# Tokenization
tiktoken==0.12.0
nltk==3.9.2

# Evaluation
#ragas==0.1.0
#datasets==2.14.0

# API and UI
#fastapi==0.104.1
#uvicorn==0.24.0
#streamlit==1.28.0

# Development
#pytest==7.4.3
#black==23.11.0
#flake8==6.1.0

# Utilities
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
